+++ CHANGELOG.md
@@ -15,6 +15,7 @@ At the moment this project **does not** adhere to
 ### Added
 - Add a way to change program modification account  ([#843](https://github.com/entropyxyz/entropy-core/pull/843))
 - Add support for `--mnemonic-file` and `THRESHOLD_SERVER_MNEMONIC` ([#864](https://github.com/entropyxyz/entropy-core/pull/864))
+- Add validator helpers to cli ([#870](https://github.com/entropyxyz/entropy-core/pull/870))
 
 ### Changed
 - Move TSS mnemonic out of keystore [#853](https://github.com/entropyxyz/entropy-core/pull/853)